requirements.txt
requests
beautifulsoup4
numpy
camelot-py
jsonschema>=4.21.0
pytest
